Adds a custom diff function for projects, to see exactly how a project changed
Several methods were added to the project model, the most relevant being
@project.changes.  The 'changes' method compares the current
object to the last saved version of that object. The result is a changes
hash where the keys are the attribute or association name as a string, and the
value is an array of length 2 where the first element is the previous value and
the second element is the current value.  If an x-to-many association was changed, then
the two values in the array will be sub arrays filled with id numbers of the
associated objects.

It's important to realize that certain unrelevant attributes are not matched:
updated_at, last_state, and iteration.

The comparison is done by adding a last_state attribute to the project model,
which is a YAML string of a the full attribute+association hash returned by the
new full_hash method on the project. If the last_state attribute is nil, then
the changes method returns an empty hash.

The save_state method updates the last_state attribute and is called
before setting the parameters in project/edit method in the Project controller.
This new functionality is added to the recent action in the static pages
controller, where it adds what changes have been made to the 'action' string of
the recent project changes.

As of now, when a new file is added to the project, the project is 'touched'
and the updated_at attribute gets bumped, but the changes method is not
able to find any differences, because files aren't saved with the model. So in
the 'recent' action, if we get a file that has been updated, but changes gives
an empty hash, that makes us think we might have new file and should
investigate.